requirements.txt
Kivy==2.3.1
databricks-sql-connector==3.4.0
